fixed chat groups auth

+++ backend/app/Http/Controllers/ChatController.php
@@ -24,21 +24,26 @@ class ChatController extends Controller
         return response()->json($chat);
     }
 
-    public function store(Request $request)
+    public function store(Request $request, Team $team)
     {
 
         Gate::authorize('create', Chat::class);
 
        $validated = $request->validate([
             'name' => 'required|string|max:255',
-            'team_id' => 'required'
         ]);
-        
 
-      $chat = Chat::create([
-      'name'=> $validated['name'],
-      'team_id' => $validated['team_id']
-      ]);
+        $user = Auth::user();
+        
+        if($team->users()->where('user_id', $user->id)){
+
+            $chat = Chat::create([
+                'name'=> $validated['name'],
+                'team_id' => $validated['team_id']
+                ]);
+        }
+
+    
         return response()->json(['success' => true, 'chat' => $chat]);
     }
 
@@ -62,9 +67,14 @@ class ChatController extends Controller
             'name' => 'required|string|max:255',
         ]);
 
+        $user = Auth::user();
+        
+        if($chat->team->users()->where('user_id', $user->id)){
+
         $chat->update([
             'name'=> $validated['name'],
         ]);
+    }
 
         return response()->json(['success' => true, 'chat' => $chat]);
     }
@@ -83,7 +93,11 @@ class ChatController extends Controller
             return response()->json(['success' => false, 'message' => 'Chat not found'], 404);
         }
 
+        if($chat->team->users()->where('user_id', $user->id)){
+
         $chat->delete();
+        }
+        
         return response()->json(['success' => true]);
     }
